Split view_utils
