+++ ydb/public/lib/ydb_cli/dump/util/view_utils.cpp
@@ -1,184 +1,19 @@
-#include "rewrite_query.h"
+#include "query_utils.h"
 #include "view_utils.h"
 
 #include <yql/essentials/parser/proto_ast/gen/v1/SQLv1Lexer.h>
 #include <yql/essentials/parser/proto_ast/gen/v1_proto_split/SQLv1Parser.pb.main.h>
-#include <yql/essentials/sql/settings/translation_settings.h>
-#include <yql/essentials/sql/v1/format/sql_format.h>
-#include <yql/essentials/sql/v1/proto_parser/proto_parser.h>
+#include <yql/essentials/public/issue/yql_issue.h>
 
-#include <library/cpp/protobuf/util/simple_reflection.h>
-
-#include <util/folder/pathsplit.h>
 #include <util/string/builder.h>
-#include <util/string/strip.h>
 
-#include <format>
 #include <re2/re2.h>
+
+namespace NYdb::NDump {
 
 using namespace NSQLv1Generated;
 
 namespace {
-
-TString RewriteAbsolutePath(TStringBuf path, TStringBuf backupRoot, TStringBuf restoreRoot) {
-    if (backupRoot == restoreRoot) {
-        return TString(path);
-    }
-
-    TPathSplitUnix pathSplit(path);
-    TPathSplitUnix backupRootSplit(backupRoot);
-
-    size_t matchedParts = 0;
-    while (matchedParts < pathSplit.size() && matchedParts < backupRootSplit.size()
-        && pathSplit[matchedParts] == backupRootSplit[matchedParts]
-    ) {
-        ++matchedParts;
-    }
-
-    TPathSplitUnix restoreRootSplit(restoreRoot);
-    for (size_t unmatchedParts = matchedParts + 1; unmatchedParts <= backupRootSplit.size(); ++unmatchedParts) {
-        restoreRootSplit.AppendComponent("..");
-    }
-    return restoreRootSplit.AppendMany(pathSplit.begin() + matchedParts, pathSplit.end()).Reconstruct();
-}
-
-struct TAbsolutePathRewriter {
-
-    static bool IsAbsolutePath(TStringBuf path) {
-        return path.StartsWith("`/") && path.EndsWith('`');
-    }
-
-    TString RewriteAbsolutePath(const TString& path) const {
-        if (BackupRoot == RestoreRoot) {
-            return path;
-        }
-
-        return TStringBuilder() << '`'
-            << ::RewriteAbsolutePath(TStringBuf(path.begin() + 1, path.end() - 1), BackupRoot, RestoreRoot)
-            << '`';
-    }
-
-    TString operator()(const TString& path) const {
-        if (IsAbsolutePath(path)) {
-            return RewriteAbsolutePath(path);
-        }
-        return path;
-    }
-
-    TStringBuf BackupRoot;
-    TStringBuf RestoreRoot;
-};
-
-struct TTokenCollector {
-    TTokenCollector(std::function<TString(const TString&)>&& pathRewriter = {}) : PathRewriter(std::move(pathRewriter)) {}
-
-    void operator()(const NProtoBuf::Message& message) {
-        if (const auto* token = dynamic_cast<const TToken*>(&message)) {
-            const auto& value = token->GetValue();
-            if (token->GetId() != NALPDefault::SQLv1LexerTokens::TOKEN_EOF) {
-                if (!Tokens.empty()) {
-                    Tokens << ' ';
-                }
-                Tokens << (IsTableRefDescendent && PathRewriter ? PathRewriter(value) : value);
-            }
-        }
-    }
-
-    TStringBuilder Tokens;
-
-    bool IsTableRefDescendent = false;
-    std::function<TString(const TString&)> PathRewriter;
-};
-
-void VisitAllFields(const NProtoBuf::Message& msg, const std::function<bool(const NProtoBuf::Message&)>& callback) {
-    const auto* descr = msg.GetDescriptor();
-    for (int i = 0; i < descr->field_count(); ++i) {
-        const auto* fd = descr->field(i);
-        NProtoBuf::TConstField field(msg, fd);
-        if (field.IsMessage()) {
-            for (size_t j = 0; j < field.Size(); ++j) {
-                const auto& message = *field.Get<NProtoBuf::Message>(j);
-                if (callback(message)) {
-                    VisitAllFields(message, callback);
-                }
-            }
-        }
-    }
-}
-
-void VisitAllFields(const NProtoBuf::Message& msg, TTokenCollector& callback) {
-    const auto* descr = msg.GetDescriptor();
-    for (int i = 0; i < descr->field_count(); ++i) {
-        const auto* fd = descr->field(i);
-        NProtoBuf::TConstField field(msg, fd);
-        if (field.IsMessage()) {
-            for (size_t j = 0; j < field.Size(); ++j) {
-                const auto& message = *field.Get<NProtoBuf::Message>(j);
-                const auto* tableRef = dynamic_cast<const TRule_table_ref*>(&message);
-                if (tableRef) {
-                    callback.IsTableRefDescendent = true;
-                }
-
-                callback(message);
-                VisitAllFields(message, callback);
-
-                if (tableRef) {
-                    callback.IsTableRefDescendent = false;
-                }
-            }
-        }
-    }
-}
-
-struct TTableRefValidator {
-
-    // returns true if the message is not a table ref and we need to dive deeper to find it
-    bool operator()(const NProtoBuf::Message& message) {
-        const auto* tableRef = dynamic_cast<const TRule_table_ref*>(&message);
-        if (!tableRef) {
-            return true;
-        }
-
-        // implementation note: a better idea might be to create a custom grammar for validation
-        if (tableRef->HasBlock3() && tableRef->GetBlock3().HasAlt1() && tableRef->GetBlock3().GetAlt1().HasRule_table_key1()) {
-            // Table keys are considered save for view backups.
-            return false;
-        }
-
-        // The only kind of table references in views that we really cannot restore are evaluated absolute paths:
-        // $path = "/old_db" || "/t"; select * from $path;
-        // If the view is being restored to a different database (like "/new_db"),
-        // then the saved create view statement will need manual patching to succeed.
-        TTokenCollector tokenCollector;
-        VisitAllFields(*tableRef, tokenCollector);
-        const TString refString = tokenCollector.Tokens;
-
-        Issues.AddIssue(TStringBuilder() << "Please check that the table reference: " << refString.Quote()
-            << " contains no evaluated expressions."
-        );
-        Issues.back().Severity = NYql::TSeverityIds::S_WARNING;
-
-        return false;
-    }
-
-    NYql::TIssues& Issues;
-};
-
-bool ValidateTableRefs(const TRule_sql_query& query, NYql::TIssues& issues) {
-    TTableRefValidator tableRefValidator(issues);
-    VisitAllFields(query, tableRefValidator);
-    return tableRefValidator.Issues.Empty();
-}
-
-TString RewriteTableRefs(const TRule_sql_query& query, TStringBuf backupRoot, TStringBuf restoreRoot) {
-    TAbsolutePathRewriter pathRewriter;
-    pathRewriter.BackupRoot = backupRoot;
-    pathRewriter.RestoreRoot = restoreRoot;
-
-    TTokenCollector tokenCollector(std::move(pathRewriter));
-    VisitAllFields(query, tokenCollector);
-    return tokenCollector.Tokens;
-}
 
 struct TViewQuerySplit {
     TString ContextRecreation;
@@ -198,28 +33,6 @@ TViewQuerySplit SplitViewQuery(TStringInput query) {
     }
 
     return split;
-}
-
-bool SqlToProtoAst(const TString& query, TRule_sql_query& queryProto, NYql::TIssues& issues) {
-    NSQLTranslation::TTranslationSettings settings;
-    if (!NSQLTranslation::ParseTranslationSettings(query, settings, issues)) {
-        return false;
-    }
-    if (settings.SyntaxVersion == 0) {
-        issues.AddIssue("cannot handle YQL syntax version 0");
-        return false;
-    }
-
-    google::protobuf::Arena arena;
-    const auto* parserProto = NSQLTranslationV1::SqlAST(
-        query, "query", issues, 0, settings.AnsiLexer, settings.Antlr4Parser, settings.TestAntlr4, &arena
-    );
-    if (!parserProto) {
-        return false;
-    }
-
-    queryProto = static_cast<const TSQLv1ParserAST&>(*parserProto).GetRule_sql_query();
-    return true;
 }
 
 bool ValidateViewQuery(const TString& query, NYql::TIssues& issues) {
@@ -245,48 +58,6 @@ void ValidateViewQuery(const TString& query, const TString& dbPath, NYql::TIssue
         }
         issues.AddIssue(std::move(restorabilityIssue));
     }
-}
-
-TString GetBackupRoot(TStringInput query) {
-    TString backupRoot;
-
-    constexpr TStringBuf targetLinePrefix = "-- backup root: \"";
-    constexpr TStringBuf discardedSuffix = "\"";
-    TString line;
-    while (query.ReadLine(line)) {
-        StripInPlace(line);
-        if (line.StartsWith(targetLinePrefix)) {
-            backupRoot = line.substr(
-                std::size(targetLinePrefix),
-                std::size(line) - std::size(targetLinePrefix) - std::size(discardedSuffix)
-            );
-            return backupRoot;
-        }
-    }
-
-    return backupRoot;
-}
-
-bool Format(const TString& query, TString& formattedQuery, NYql::TIssues& issues) {
-    google::protobuf::Arena arena;
-    NSQLTranslation::TTranslationSettings settings;
-    settings.Arena = &arena;
-
-    auto formatter = NSQLFormat::MakeSqlFormatter(settings);
-    return formatter->Format(query, formattedQuery, issues);
-}
-
-bool RewriteTableRefs(TString& query, TStringBuf backupRoot, TStringBuf restoreRoot, NYql::TIssues& issues) {
-    TRule_sql_query queryProto;
-    if (!SqlToProtoAst(query, queryProto, issues)) {
-        return false;
-    }
-    const auto rewrittenQuery = RewriteTableRefs(queryProto, backupRoot, restoreRoot);
-    // formatting here is necessary for the view to have pretty text inside it after the creation
-    if (!Format(rewrittenQuery, query, issues)) {
-        return false;
-    }
-    return true;
 }
 
 bool RewriteTablePathPrefix(TString& query, TStringBuf backupRoot, TStringBuf restoreRoot,
@@ -331,9 +102,7 @@ bool RewriteTablePathPrefix(TString& query, TStringBuf backupRoot, TStringBuf re
     return true;
 }
 
-}
-
-namespace NYdb::NDump {
+} // anonymous
 
 TString BuildCreateViewQuery(
     const TString& name, const TString& dbPath, const TString& viewQuery, const TString& backupRoot,
@@ -377,4 +146,4 @@ bool RewriteCreateViewQuery(TString& query, const TString& restoreRoot, bool res
     return RewriteCreateQuery(query, "CREATE VIEW IF NOT EXISTS `{}`", dbPath, issues);
 }
 
-}
+} // NYdb::NDump
