+++ ydb/public/lib/ydb_cli/dump/util/view_utils.h
@@ -1,4 +1,10 @@
-#include <yql/essentials/public/issue/yql_issue.h>
+#pragma once
+
+#include <util/generic/string.h>
+
+namespace NYql {
+    class TIssues;
+}
 
 namespace NYdb::NDump {
